third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/container-name-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'container-name' to the 'my-container' keyword: my-container
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] 'container-type' does not support 'name1 name2'
  assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object Undefined]"
Harness: the test ran to completion.
